HW-4 BaseFragment생성 및 상속 구조 변경
